% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{autoTextAnalysis}
\alias{autoTextAnalysis}
\title{Cleans up text and perfrom basic text analysis}
\usage{
  textAnalysisFunction(input)
}
\arguments{
  \item{input}{input data set}
}
\description{
  This package accepts txt files, and after cleaning them up
  (removing stop words, punctucations, etc), conducts basic semantic analytics on them. 
  These analytics functions include : 1. creating a word cloud to highlight the most important words
  2.Utilizing lexicon_bing dictionary to give a ration of negative and positive words to the user as well 
  as a char that shows that ratio.
}
